Refactor: Update Android workflow artifact name and enhance IslandOverlayView with launcher apps integration
