+++ .github/workflows/android.yml
@@ -71,7 +71,7 @@ jobs:
         if: ${{ github.event_name == 'pull_request' || github.ref == 'refs/heads/dev' }}
         uses: actions/upload-artifact@v4
         with:
-          name: SmartIsland-debug-${{ github.run_attempt }}
+          name: SmartIsland-Debug-apk
           path: app/build/outputs/apk/debug/app-debug.apk
           if-no-files-found: error
           retention-days: 14
