+++ app/src/main/java/com/agupta07505/smartisland/ui/OverlayIsland.kt
@@ -10,8 +10,15 @@ package com.agupta07505.smartisland.ui
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.produceState
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import com.agupta07505.smartisland.data.AppShortcutProvider
+import com.agupta07505.smartisland.data.LaunchableApp
 import com.agupta07505.smartisland.model.IslandNotification
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 @Composable
 fun OverlayIsland(
@@ -26,12 +33,35 @@ fun OverlayIsland(
     val expanded by viewModel.expanded.collectAsState()
     val notifications by viewModel.notifications.collectAsState()
     val selectedIndex by viewModel.selectedIndex.collectAsState()
+    val context = LocalContext.current
+    val selectedApps = remember(settings.shortcutPackages) {
+        AppShortcutProvider.selectedApps(context, settings.shortcutPackages)
+    }
+    val launcherApps by produceState<List<LaunchableApp>?>(
+        initialValue = when {
+            selectedApps.isNotEmpty() -> selectedApps
+            settings.shortcutPackages.isEmpty() && !settings.showRecentApps -> emptyList()
+            !AppShortcutProvider.hasUsageAccess(context) -> emptyList()
+            else -> null
+        },
+        settings.shortcutPackages,
+        settings.showRecentApps
+    ) {
+        value = withContext(Dispatchers.IO) {
+            AppShortcutProvider.shortcuts(
+                context = context,
+                selectedPackages = settings.shortcutPackages,
+                includeRecent = settings.showRecentApps
+            )
+        }
+    }
 
     IslandOverlayView(
         settings = settings,
         expanded = expanded,
         notifications = notifications,
         selectedIndex = selectedIndex,
+        launcherApps = launcherApps,
         onPageSelected = { index -> viewModel.setSelectedNotificationIndex(index) },
         onOpenNotification = onOpenNotification,
         onLaunchApp = onLaunchApp,
